+++ src/components/app/App.jsx
@@ -6,10 +6,11 @@ import { get } from 'lodash';
 import { isFHIRServer, isLoggedIn } from '../../common/utils';
 import Search from '../search/Search';
 import ConceptHome from '../concepts/ConceptHome';
-import ConceptsComparison from '../concepts/ConceptsComparison';
 import MappingHome from '../mappings/MappingHome';
 import SourceHome from '../sources/SourceHome';
 import CollectionHome from '../collections/CollectionHome';
+import ConceptsComparison from '../concepts/ConceptsComparison';
+import MappingsComparison from '../mappings/MappingsComparison';
 import OrgHome from '../orgs/OrgHome';
 import UserHome from '../users/UserHome';
 import Login from '../users/Login';
@@ -28,6 +29,7 @@ import ConceptMapHome from '../fhir/ConceptMapHome';
 import Header from './Header';
 import Footer from './Footer';
 import RootView from './RootView';
+import DocumentTitle from "./DocumentTitle"
 import './App.scss';
 
 
@@ -66,6 +68,7 @@ const App = props => {
 
   return (
     <div>
+      <DocumentTitle/>
       <Header {...props} onOpen={setMenuOpen} />
       <ErrorBoundary>
         <main className={menuOpen ? 'content menu-open' : 'content'}>
@@ -106,11 +109,6 @@ const App = props => {
             <Route
               path="/orgs/:org([a-zA-Z0-9\-\.\_\@]+)/sources/:source([a-zA-Z0-9\-\.\_\@]+)/:version([a-zA-Z0-9\-\.\_\@]+)/concepts/:concept"
               component={ConceptHome}
-            />
-            <Route
-              exact
-              path="/concepts/compare"
-              component={ConceptsComparison}
             />
 
             { /* Mapping Home */ }
@@ -180,6 +178,18 @@ const App = props => {
             <Route
               path="/users/:user([a-zA-Z0-9\-\.\_\@]+)/collections/:collection([a-zA-Z0-9\-\.\_\@]+)"
               component={CollectionHome}
+            />
+
+            {/* Comparison */}
+            <Route
+              exact
+              path="/concepts/compare"
+              component={ConceptsComparison}
+              />
+            <Route
+              exact
+              path="/mappings/compare"
+              component={MappingsComparison}
             />
 
             {/* Organization Home */}