+++ src/components/search/Search.jsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { withRouter } from "react-router";
+import alertifyjs from 'alertifyjs';
 import {
   get, set, cloneDeep, merge, forEach, includes, keys, pickBy, size, isEmpty, has, find, isEqual,
   map, omit, isString
@@ -24,7 +25,7 @@ import SearchByAttributeInput from './SearchByAttributeInput';
 import ResourcesHorizontal from './ResourcesHorizontal';
 import ResourceTabs from './ResourceTabs';
 //import Resources from './Resources';
-import { fetchSearchResults, fetchCounts } from './utils';
+import { fetchSearchResults, fetchCounts, fetchFacets } from './utils';
 import LayoutToggle from '../common/LayoutToggle';
 import InfiniteScrollChip from '../common/InfiniteScrollChip';
 import { FACET_ORDER } from './ResultConstants';
@@ -70,6 +71,7 @@ class Search extends React.Component {
       includeRetired: false,
       userFilters: {},
       isURLUpdatedByActionChange: false,
+      facets: {},
       results: {
         concepts: cloneDeep(resourceResultStruct),
         mappings: cloneDeep(resourceResultStruct),
@@ -229,8 +231,7 @@ class Search extends React.Component {
     const numReturned = parseInt(get(response, 'headers.num_returned', 0))
     const next = get(response, 'headers.next')
     const previous = get(response, 'headers.previous')
-    const facets = get(response, 'data.facets', {})
-    let items = get(response, 'data.results', [])
+    let items = get(response, 'data', [])
     if(this.state.isInfinite && !resetItems)
       items = [...this.state.results[resource].items, ...items]
     return {
@@ -240,7 +241,7 @@ class Search extends React.Component {
       pages: parseInt(get(response, 'headers.pages', 1)),
       next: next,
       prev: previous,
-      facets: facets,
+      facets: this.state.facets,
       items: items,
     }
   }
@@ -257,9 +258,20 @@ class Search extends React.Component {
         if(includes(['sources', 'collections'], resource))
           this.loadSummary(resource)
       })
-    } else {
-      this.setState({isLoading: false}, () => {
-        throw response
+    } else if (get(response, 'detail'))
+      this.setState({isLoading: false}, () => alertifyjs.error(response.detail, 0))
+    else
+      this.setState({isLoading: false}, () => {throw response})
+  }
+
+  onFacetsLoad = (response, resource) => {
+    if(response.status === 200) {
+      this.setState({facets: get(response, 'data.facets', {})}, () => {
+        if(has(this.state, `results.${resource}.facets`)) {
+          const newState = {...this.state}
+          newState.results[resource].facets = this.state.facets
+          this.setState(newState)
+        }
       })
     }
   }
@@ -301,7 +313,7 @@ class Search extends React.Component {
     return {...queryParam, ...viewFilters}
   }
 
-  fetchNewResults(attrsToSet, counts=true, resetItems=true, updateURL=false) {
+  fetchNewResults(attrsToSet, counts=true, resetItems=true, updateURL=false, facets=true) {
     if(!attrsToSet)
       attrsToSet = {}
 
@@ -318,6 +330,7 @@ class Search extends React.Component {
     }
     if(!this.props.fhir)
       newState.isURLUpdatedByActionChange = true
+    newState.facets = {}
     this.setState(newState, () => {
       const {
         resource, searchStr, page, exactMatch, sortParams, updatedSince, limit,
@@ -351,7 +364,6 @@ class Search extends React.Component {
       fetchSearchResults(
         _resource,
         params,
-        !noHeaders,
         this.props.baseURL,
         null,
         response => {
@@ -363,11 +375,13 @@ class Search extends React.Component {
       )
       if(counts && !this.props.nested)
         fetchCounts(_resource, queryParams, this.onCountsLoad)
+      if(!noHeaders && facets && !fhir && resource !== 'references')
+        fetchFacets(_resource, queryParams, this.props.baseURL, this.onFacetsLoad)
     })
   }
 
   loadMore = () => {
-    this.fetchNewResults({page: this.state.page + 1}, false, false, true);
+    this.fetchNewResults({page: this.state.page + 1}, false, false, true, false);
   }
 
   onPageChange = page => {
@@ -381,7 +395,7 @@ class Search extends React.Component {
           }
         }, () => this.fetchNewResults(null, false, false))
       else
-        this.fetchNewResults({page: page}, false, false, true)
+        this.fetchNewResults({page: page}, false, false, true, false)
 
     }
   }
@@ -395,7 +409,7 @@ class Search extends React.Component {
       }, () => this.fetchNewResults(null, false, true))
     }
     else
-      this.setState({sortParams: params}, () => this.fetchNewResults(null, false, true, true))
+      this.setState({sortParams: params}, () => this.fetchNewResults(null, false, true, true, false))
   }
 
   hasPrev() {
@@ -441,7 +455,7 @@ class Search extends React.Component {
     return 'All Time'
   }
 
-  onClickIncludeRetired = () => this.fetchNewResults({includeRetired: !this.state.includeRetired}, true, true, true)
+  onClickIncludeRetired = () => this.fetchNewResults({includeRetired: !this.state.includeRetired}, true, true, true, false)
 
   onLayoutChange = newLayoutId => {
     const existingLayoutId = this.getLayoutTypeName()
@@ -627,14 +641,14 @@ class Search extends React.Component {
 
   onLimitChange = limit => this.props.fhir ?
                          this.setState({limit: limit, fhirParams: {...this.state.fhirParams, _count: limit, _getpagesoffset: 0}}, () => this.fetchNewResults(null, false, false)) :
-                         this.fetchNewResults({limit: limit}, false, true, true)
+                         this.fetchNewResults({limit: limit}, false, true, true, false)
 
   toggleFacetsDrawer = () => this.setState({openFacetsDrawer: !this.state.openFacetsDrawer})
 
   onCloseFacetsDrawer = () => this.setState({openFacetsDrawer: false})
 
   onApplyFacets = filters => this.setState(
-    {appliedFacets: filters}, () => this.fetchNewResults(null, false, true)
+    {appliedFacets: filters}, () => this.fetchNewResults(null, false, true, false, false)
   )
 
   onApplyUserFilters = (id, value) => {